# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: E-Procurement System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deactivateUser
      tags: [users]
      summary: Deactivate user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    User:
      type: object
      required: [id, username, fullName, role, department, approvalStage, isActive, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        role:
          type: string
        department:
          type: string
        approvalStage:
          type: ["integer", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    UserInput:
      type: object
      required: [username, password, fullName, role, department]
      properties:
        username:
          type: string
        password:
          type: string
        fullName:
          type: string
        role:
          type: string
        department:
          type: string
        approvalStage:
          type: ["integer", "null"]
    UserUpdate:
      type: object
      properties:
        fullName:
          type: string
        role:
          type: string
        department:
          type: string
        approvalStage:
          type: ["integer", "null"]
        isActive:
          type: boolean
        password:
          type: string
...