# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * E-Procurement System API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  fullName?: string;
  role?: string;
  department?: string;
  /** @nullable */
  approvalStage?: number | null;
  isActive?: boolean;
  password?: string;
}